LayoutTests/http/tests/media/fairplay/fps-init-data-skd-expected.txt

PROMISE: requestMediaKeySystemAccess resolved
PROMISE: createMediaKeys resolved
FETCH: server certificate recieved
PROMISE: arrayBuffer resolved
PROMISE: setServerCertificate resolved
EVENT(message)
PROMISE: licenseResponse resolved
PROMISE: session.update() resolved
END OF TEST
